Update angular-locale_es-es.js
The language codes for months and days in Spanish translations are misspelled. They look like "ene.", "feb.", "lun.", "mar.", and so on, but in Spanish grammar that final point is not applied to date codes, just to abbreviations ("en.", "febr.", "lun.", "mart.", etc.).

They must be changed in all "es-xx" JS files.